cleaned up the directories
